Move chart form deep explore to components
