+++ app/components/EntityCreationLineChart/ResolutionSelectInput/index.tsx
